suscripciones_app/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
google-api-python-client==2.117.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
